Permisos modulo Usuario

+++ views/Profesor/ProfesorCon.php
@@ -61,20 +61,26 @@ require_once "./core/configGeneral.php";
 
             <div class="titles">
                 <h1 class="tituloa ">Profesores</h1>
-                <div class="form-group" id="botones">
-                    <!--Pará mostrar una línea horizontal en el documento-->
-                    <a href="?c=Profesor&a=nuevo" class="boton btn btn-info botones">Nuevo Profesor</a>
-
-                    <!--Pará crear un contenedor adaptable con los estilos de bootstrap-->
-
-
-                </div>
+                <?php
+                    if($sesion=='A'){
+                        echo '                
+                        <div class="form-group" id="botones">
+                        <!--Pará mostrar una línea horizontal en el documento-->
+                        <a href="?c=Profesor&a=nuevo" class="boton btn btn-info botones">Nuevo Profesor</a>
+    
+                        <!--Pará crear un contenedor adaptable con los estilos de bootstrap-->
+    
+    
+                    </div>';}
+                    else{echo ' ';}
+                    
+                ?>
             </div>
 
             <div class="tablas">
             <table id="table" data-toggle="table" data-height="auto" data-search="true" data-show-refresh="false"
                 data-show-columns="true" data-buttons-toolbar=".buttons-toolbar">
-                <thead">
+                <thead>
                     <tr>
                         <th data-field="ID">ID</th>
                         <th data-field="nombre">Nombre</th>
@@ -83,7 +89,11 @@ require_once "./core/configGeneral.php";
                         <th data-field="Doc">Documento</th>
                         <th data-field="especialidad">Especialidad</th>
                         <th data-field="editar"></th>
-                        <th data-field="eliminar"></th>
+                        <?php
+                            if($sesion=='A'){
+                                echo '<th data-field="eliminar"></th>';
+                            }
+                        ?>
                     </tr>
                 </thead>
                 <tbody>
@@ -106,19 +116,23 @@ require_once "./core/configGeneral.php";
 
                             </div>
                         </th>
-                        <th data-field="eliminar">
-                        <div class="form-group" id="botones">
-                                <!--Pará mostrar una línea horizontal en el documento-->
-                                <a onclick="javascript:return confirm('¿Desea eliminar el registro?');
-                                " href="?c=Profesor&a=g_eliminar_Profesor&id_Usuario=<?php echo $k->id_Usuario;?>"
-                                 class="fas fa-trash-alt fa-lg"></a>
-                               
-
-                                <!--Pará crear un contenedor adaptable con los estilos de bootstrap-->
-
-
-                            </div>
-                        </th>
+                        <?php
+                            if($sesion=='A'){
+                                echo '<th data-field="eliminar">
+                                <div class="form-group" id="botones">
+                                        <!--Pará mostrar una línea horizontal en el documento-->
+                                        <a onclick="javascript:return confirm(\'¿Desea eliminar el registro?\');
+                                        " href="?c=Profesor&a=g_eliminar_Profesor&id_Usuario=<?php echo $k->id_Usuario;?>"
+                                         class="fas fa-trash-alt fa-lg"></a>
+                                       
+        
+                                        <!--Pará crear un contenedor adaptable con los estilos de bootstrap-->
+        
+        
+                                    </div>
+                                </th>';
+                            }
+                        ?>
                     </tr>
                     <?php endforeach; ?>
                 </tbody>
